feat: add unique constraint to follows table; implement checks in follow and unfollow services to prevent duplicate follows

+++ src/app/modules/follow/follow.service.ts
@@ -1,20 +1,59 @@
 import { Follow } from "@prisma/client";
 import prisma from "../../config/prisma";
+import AppError from "../../errors/AppError";
+import httpStatus from "http-status";
 
-const followShop = async (payload: Follow) => {
-  const result = await prisma.follow.create({
-    data: payload,
-  });
-  return result;
-};
-const unfollowShop = async (payload: Follow) => {
-  const result = await prisma.follow.delete({
+const followShop = async (payload: Omit<Follow, "id" | "createdAt" | "updatedAt">) => {
+  const isExist = await prisma.follow.findUnique({
     where: {
-      id: `${payload.customerId}_${payload.shopId}`,
+      customerId_shopId: {
+        customerId: payload.customerId,
+        shopId: payload.shopId,
+      },
     },
   });
+
+  if (isExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, "You have already followed this shop.");
+  }
+
+  const result = await prisma.follow.create({
+    data: {
+      customerId: payload.customerId,
+      shopId: payload.shopId,
+    },
+  });
+
   return result;
 };
+
+const unfollowShop = async (payload: Omit<Follow, "id" | "createdAt" | "updatedAt">) => {
+  const isExist = await prisma.follow.findUnique({
+    where: {
+      customerId_shopId: {
+        customerId: payload.customerId,
+        shopId: payload.shopId,
+      },
+    },
+  });
+
+  if (!isExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, "You are not following this shop.");
+  }
+
+  const result = await prisma.follow.delete({
+    where: {
+      customerId_shopId: {
+        customerId: payload.customerId,
+        shopId: payload.shopId,
+      },
+    },
+  });
+
+  return result;
+};
+
+
 const getFollowedShops = async (userEmail: string) => {
   const result = await prisma.customer.findFirst({
     where: {
